src/requirements.txt
absl-py==1.3.0
addict==2.4.0
brotlipy==0.7.0
cachetools==5.2.0
certifi
cffi
charset-normalizer
contextlib2==21.6.0
contourpy==1.0.6
cryptography
cycler==0.11.0
einops==0.6.0
filelock==3.8.2
flit_core
fonttools==4.38.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.11.1
idna
importlib-metadata==5.1.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
ml-collections==0.1.1
mmcv==1.7.0
monai==1.0.1
numpy
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.2.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.1
SimpleITK==2.2.1
six
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.6.12
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions
urllib3
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.11.0
scikit-image
medpy